dam-mcp/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "dam-mcp"
version = "0.1.0"
description = "Digital Asset Management MCP server — search, upload, tag, and serve creative assets from GCS"
readme = "README.md"
requires-python = ">=3.10"
keywords = ["dam", "assets", "gcs", "mcp", "claude"]
classifiers = [
    "Programming Language :: Python :: 3",
    "Operating System :: OS Independent",
]

dependencies = [
    "mcp[cli]>=1.12.0",
    "google-cloud-storage>=2.0.0",
    "google-auth>=2.0.0",
    "httpx>=0.26.0",
    "python-dotenv>=1.1.0",
    "pytest>=8.4.1",
    "pytest-asyncio>=1.0.0",
]

[project.scripts]
dam-mcp = "dam_mcp:entrypoint"

[tool.hatch.build.targets.wheel]
packages = ["dam_mcp"]

[tool.pytest.ini_options]
markers = [
    "e2e: marks tests as end-to-end (requires running MCP server and GCS bucket) - excluded from default runs",
]
addopts = "-v --strict-markers -m 'not e2e'"
testpaths = ["tests"]
asyncio_mode = "auto"
